requirements.txt
pandas==2.3.3
pymongo==3.11.0
python-dotenv==1.0.0
schedule==1.2.0